Support configuring additional labels on resources and add ServiceAccount
Fixes https://github.com/qdrant/qdrant-helm/issues/55

This adds a ServiceAccount for the Statefulset. This errors if a namespace has no default ServiceAccount and prevents accidentally using a default namespace ServiceAccount with elevated permissions.

This also introduces terratest as a way to do unit tests on the rendered helm chart.