Make the qdrant api key easily configurable
This adds a new `apiKey` value.

If `true` it generates a random apiKey, stores it in a secret and ensures that it stays the same during upgrades.
If a string is passed, this string is used as the api key.

The secret is mounted as a `local.yaml` config into the container. See also https://qdrant.tech/documentation/guides/configuration/#order-and-priority.

Alternatively, there are two new values `additionalVolumes` and `additionalVolumeMounts` which allow to mount additional existing volumes, such as pre-existing secrets into the qdrant container.

This also adds bats-core based integration tests to test the correct api key handling

Signed-off-by: Bastian Hofmann <mail@bastianhofmann.de>