Apply tpl() to affinity values to enable reuse of helpers / existing labels
The Helm tpl function [1] allows to use templates in values. This does make
sense where values might want reference existing labels of a Helm chart.
This is is the case for Pod affinity (or anti-affinity) which needs to match
the labels of the Pods created.

This is somewhat inspired by the way this was done for the NGINX Ingress
Controller with [2].

[1] https://helm.sh/docs/howto/charts_tips_and_tricks/#using-the-tpl-function
[2] https://github.com/kubernetes/ingress-nginx/commit/af9e5246ad09a304ef68d11f1467185dbd12fe72

Resolves: #212
Signed-off-by: Christian Rohmann <christian.rohmann@inovex.de>