fix(chart): add service.enable_tls to values
The test-db-interaction test was failing because it was checking for `config.service.enable_tls` which did not exist in the `values.yaml` file.

This commit adds the `service.enable_tls` key to the `config` section of the `values.yaml` file and sets its default value to `false`. This allows the test to correctly determine whether to use http or https when connecting to the qdrant service.